shm/rma: bypass Yaksa when Get/Put with contiguous datatype
MPIR_CVAR_CH4_IPC_GPU_RMA_ENGINE_TYPE is added to allow for tuning how
GPU buffer copying is handled for contiguous data transfers in shm RMA
paths. By default, automatic selection occurs, which bypasses yaksa and
determines which GPU engine type will perform best in the active
situation.

To make the selection, the source and destination buffer placements are
examined to choose between the high-bandwidth and low-latency engines
for the active copy. The global dev_id of the window is added to
shared_table during win creation to reduce device id querying for this
check.