ROMIO: use user buffer to send if can, to reduce number of memcpy calls
ADIOI_LUSTRE_Fill_send_buffer can be expensive when the number of
noncontiguous segments is large. The write contents of noncontiguous
segments are copied into an internally allocated send buffer, a
contiguous space, which is then used in calls to MPI_Issend().

Note each collective write is performed in multiple rounds of 2-phase
I/O. In some rounds, the part of user buffer may be contiguous. This
commit checks whether or not the part of user buffer is contiguous and
used in MPI_Issend directly, if it is.

Test case: test/pio_noncontig.c