ROMIO: when # compute nodes < # OSTs, adjust striping size in collective write
For collective write, when the number of OSTs is twice or more larger
then the number of compute nodes, it can be good to have each I/O
aggregator to write to two or more consecutive OSTs. This commit will
not introduce any file lock conflict, as it still yields the effect of
any one OST receiving write requests from aggregators running on only
one compute node.